# ENH: stats.mannwhitneyu: replace exact p-value calculation