--- a/scipy/stats/_mannwhitneyu.py
+++ b/scipy/stats/_mannwhitneyu.py
@@ -121,17 +121,17 @@ class _MWU:
         s_array = self.build_sigma_array(maxu)
 
         # Start working with ints, for maximum precision and efficiency:
-        configurations = np.zeros(maxu + 1, dtype=np.int64)
+        configurations = np.zeros(maxu + 1, dtype=np.uint64)
         # How many ways to have U=0? 1
         configurations[0] = 1
 
-        with np.errstate(over='raise'):
+        with np.errstate(invalid='raise', over='raise'):
             for u in np.arange(1, maxu + 1):
                 coeffs = s_array[u - 1::-1]
                 new_val = np.dot(configurations[:u], coeffs) / u
                 try:
                     configurations[u] = new_val
-                except FloatingPointError:
+                except (FloatingPointError, OverflowError):
                     # OK, we got into numbers too big for uint64.
                     # So now we start working with floats.
                     # By doing this since the beginning, we would have lost precision.