--- a/scipy/stats/tests/test_hypotests.py
+++ b/scipy/stats/tests/test_hypotests.py
@@ -162,8 +162,6 @@ class TestCvm:
 
 
 class TestMannWhitneyU:
-    def setup_method(self):
-        _mwu_state._recursive = True
 
     # All magic numbers are from R wilcox.test unless otherwise specified
     # https://rdrr.io/r/stats/wilcox.test.html
@@ -359,21 +357,23 @@ class TestMannWhitneyU:
             for m, p in table.items():
                 # check p-value against table
                 u = np.arange(0, len(p))
-                assert_allclose(_mwu_state.cdf(k=u, m=m, n=n), p, atol=1e-3)
+                _mwu_state.set_shapes(m, n)
+                assert_allclose(_mwu_state.cdf(k=u), p, atol=1e-3)
 
                 # check identity CDF + SF - PMF = 1
                 # ( In this implementation, SF(U) includes PMF(U) )
                 u2 = np.arange(0, m*n+1)
-                assert_allclose(_mwu_state.cdf(k=u2, m=m, n=n)
-                                + _mwu_state.sf(k=u2, m=m, n=n)
-                                - _mwu_state.pmf(k=u2, m=m, n=n), 1)
+                assert_allclose(_mwu_state.cdf(k=u2)
+                                + _mwu_state.sf(k=u2)
+                                - _mwu_state.pmf(k=u2), 1)
 
                 # check symmetry about mean of U, i.e. pmf(U) = pmf(m*n-U)
-                pmf = _mwu_state.pmf(k=u2, m=m, n=n)
+                pmf = _mwu_state.pmf(k=u2)
                 assert_allclose(pmf, pmf[::-1])
 
                 # check symmetry w.r.t. interchange of m, n
-                pmf2 = _mwu_state.pmf(k=u2, m=n, n=m)
+                _mwu_state.set_shapes(n, m)
+                pmf2 = _mwu_state.pmf(k=u2)
                 assert_allclose(pmf, pmf2)
 
     def test_asymptotic_behavior(self):
@@ -615,24 +615,25 @@ class TestMannWhitneyU:
         # sizes n1 and n2 changed. This was indicative of unnecessary cache
         # growth and redundant calculation. Check that this is resolved.
         rng = np.random.default_rng(7600451795963068007)
-        x = rng.random(size=5)
-        y = rng.random(size=11)
-        _mwu_state._fmnks = -np.ones((1, 1, 1))  # reset cache
-        stats.mannwhitneyu(x, y, method='exact')
-        shape = _mwu_state._fmnks.shape
-        assert shape[0] <= 6 and shape[1] <= 12  # one more than sizes
+        m, n = 5, 11
+        x = rng.random(size=m)
+        y = rng.random(size=n)
+        _mwu_state.reset()  # reset cache
+        res = stats.mannwhitneyu(x, y, method='exact')
+        shape = _mwu_state.configurations.shape
+        assert shape[-1] == min(res.statistic, m*n - res.statistic) + 1
         stats.mannwhitneyu(y, x, method='exact')
-        assert shape == _mwu_state._fmnks.shape  # unchanged when sizes are reversed
+        assert shape == _mwu_state.configurations.shape  # unchanged when sizes are reversed
 
         # Also, we weren't exploiting the symmmetry of the null distribution
         # to its full potential. Ensure that the null distribution is not
         # evaluated explicitly for `k > m*n/2`.
-        _mwu_state._fmnks = -np.ones((1, 1, 1))  # reset cache
+        _mwu_state.reset()  # reset cache
         stats.mannwhitneyu(x, 0*y, method='exact', alternative='greater')
-        shape = _mwu_state._fmnks.shape
+        shape = _mwu_state.configurations.shape
         assert shape[-1] == 1  # k is smallest possible
         stats.mannwhitneyu(0*x, y, method='exact', alternative='greater')
-        assert shape == _mwu_state._fmnks.shape
+        assert shape == _mwu_state.configurations.shape
 
     @pytest.mark.parametrize('alternative', ['less', 'greater', 'two-sided'])
     def test_permutation_method(self, alternative):
@@ -645,43 +646,6 @@ class TestMannWhitneyU:
                                   alternative=alternative, axis=1)
         assert_allclose(res.statistic, res2.statistic, rtol=1e-15)
         assert_allclose(res.pvalue, res2.pvalue, rtol=1e-15)
-
-    def teardown_method(self):
-        _mwu_state._recursive = None
-
-
-class TestMannWhitneyU_iterative(TestMannWhitneyU):
-    def setup_method(self):
-        _mwu_state._recursive = False
-
-    def teardown_method(self):
-        _mwu_state._recursive = None
-
-
-@pytest.mark.xslow
-def test_mann_whitney_u_switch():
-    # Check that mannwhiteneyu switches between recursive and iterative
-    # implementations at n = 500
-
-    # ensure that recursion is not enforced
-    _mwu_state._recursive = None
-    _mwu_state._fmnks = -np.ones((1, 1, 1))
-
-    rng = np.random.default_rng(9546146887652)
-    x = rng.random(5)
-
-    # use iterative algorithm because n > 500
-    y = rng.random(501)
-    stats.mannwhitneyu(x, y, method='exact')
-    # iterative algorithm doesn't modify _mwu_state._fmnks
-    assert np.all(_mwu_state._fmnks == -1)
-
-    # use recursive algorithm because n <= 500
-    y = rng.random(500)
-    stats.mannwhitneyu(x, y, method='exact')
-
-    # recursive algorithm has modified _mwu_state._fmnks
-    assert not np.all(_mwu_state._fmnks == -1)
 
 
 class TestSomersD(_TestPythranFunc):