#### Reference issue
gh-4933
Closes gh-17559

#### What does this implement/fix?
https://github.com/scipy/scipy/pull/4933#issuecomment-1898082691 suggested a different algorithm for the `mannwhitneyu` exact p-value calculation that greatly improves performance (e.g. factor of 675 for $m \approx n \approx 100$) and reduces memory consumption from $O(m^2n^2)$ to $O(mn)$. This PR implements the suggestion.

#### Additional information
@AtsushiSakai let's wait for confirmation from @toobaz on use of the code before starting a detailed review. In the meantime, I thought you'd be interested in this:

```python3
import numpy as np
from scipy import stats
from scipy.stats._mannwhitneyu import _mwu_state
rng = np.random.default_rng(1305672164505329978)
x = rng.random(100)
y = rng.random(101)
```

Before:
```python3
res = stats.mannwhitneyu(x, y, method='exact')  # 16.9 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
res  # MannwhitneyuResult(statistic=4792.0, pvalue=0.5331021234219595)
_mwu_state._fmnks.shape  #  (101, 102, 4793)
```

After:
```python3
res = stats.mannwhitneyu(x, y, method='exact')  # 25 ms ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
res  # MannwhitneyuResult(statistic=4792.0, pvalue=0.5331021234219553)
_mwu_state.configurations.shape  #  (4793,)
```