@mdhaber 2024-02-10
I rarely consider review-by-test to be acceptable, but in this case I do. So if you're comfortable with it, feel free to merge.