@AtsushiSakai 2024-02-10
Give me a week or so to test the new implementation using my data collections. Then I will merge it.