@toobaz 2024-02-10
Apologies for not finding the time before to seriously look at this, but when I wrote this Python code I had also written a Cython version, and a few quick tests I just ran show a speed up between 4x and 5x asymptoticaly (and even larger for small values of U).

If there's interest, I should be able to put up a PR next week (in this case, do let me know if there is anything specific I must keep in mind to contribute cython code to scipy).