@mdhaber 2024-02-10
I would prefer to use Pythran. This is already a massive speedup, so i think we can merge this as is and move the loop into Pythran later. Thanks, though!