@toobaz 2024-02-11
> I would prefer to use Pythran. This is already a massive speedup, so i think we can merge this as is and move the loop into Pythran later.

OK. Just a double check: this code uses `scipy.special.comb`, written in cython, and hence to the best of my understanding not trivial to use from Pythran. (But the best of my understanding is really not much)