@steppi 2024-02-11
> OK. Just a double check: this code uses `scipy.special.comb`, written in cython, and hence to the best of my understanding not trivial to use from Pythran. (But the best of my understanding is really not much)

`binom` is supported in Pythran so it should be fine (https://pythran.readthedocs.io/en/latest/SUPPORT.html#special).