@AtsushiSakai 2024-03-03
@mdhaber Sorry, I've tried compiling this PR branch on my MacBook environment several times, but the build for scipy is not passing. It seems to be the same issue related to clang as mentioned here, #18645
but I haven't been able to resolve it. Therefore, I would like to ask someone else to review it. I'm very sorry.