@lucascolley 2024-03-03
Hmm, our `cpp_std` is set to `17` in the top level meson file now so that specific issue shouldn't be here anymore... would you mind posting the error message over there or in a new issue?