@mdhaber 2024-03-12
@dschmitz89 It would be great to get this in 1.13.0 since it has huge speed and memory benefits for `mannwhitneyu`. Do you have time before branch to take a look? I don't even think it needs any stats-specific review; I'm treating the code as vendored and the tests that it is producing correct p-values are already pretty solid.