@mdhaber 2024-03-24
Thanks @AtsushiSakai. To provide further evidence of the improvement, I used the `_mwu_state` object to compute all values of the survival function (that is, at all values within the support of the distribution) for cases in which `m = n`. The plot below shows the execution time (on a log scale) for each value of `m = n`.

<img width="600" alt="image" src="https://github.com/scipy/scipy/assets/6570539/98c43f3f-2fe9-4a86-83d5-7e3149ebf194">

All values tested matched to within a relative tolerance of `5e-14`. I checked with `n = 2*m`, and observations are similar. As mentioned in the PR description, it was clear analytically that the memory usage is reduced from $O(m^2n^2)$ to $O(mn)$.

<details>

```python3
import numpy as np
import matplotlib.pyplot as plt
from time import perf_counter
from scipy.stats._mannwhitneyu import _mwu_state

mn = np.arange(10, 100)

def benchmark_pr(mn):
    times = np.empty(mn.shape, dtype=np.float64)
    ps = {}
    for i in mn:
        _mwu_state.reset()
        m = i
        n = m
        u = np.arange(m*n + 1)
        _mwu_state.set_shapes(m, n)

        tic = perf_counter()
        p = _mwu_state.sf(k=u)
        toc = perf_counter()

        times[i - mn[0]] = toc-tic
        ps[i] = p
    return times, ps

def benchmark_main(mn):
    times = np.empty(mn.shape, dtype=np.float64)
    ps = {}
    for i in mn:
        _mwu_state._fmnks = -np.ones((1, 1, 1))
        m = i
        n = m
        u = np.arange(m*n + 1)

        tic = perf_counter()
        p = _mwu_state.sf(k=u, m=m, n=n)
        toc = perf_counter()

        times[i - mn[0]] = toc-tic
        ps[i] = p
    return times, ps

# Run this with PR
times, ps = benchmark_pr(mn)
np.savez('mwu_pr.npz', mn=mn, times=times, ps=ps)
plt.semilogy(mn, times)
plt.show()

# # Run this with main
# times, ps = benchmark_main(mn)
# np.savez('mwu_main.npz', mn=mn, times=times, ps=ps)
# plt.semilogy(mn, times)
# plt.show()

# # Then execute this
# res_pr = np.load('mwu_pr.npz', allow_pickle=True)
# mn_pr, times_pr, ps_pr = res_pr['mn'], res_pr['times'], res_pr['ps']
#
# res_main = np.load('mwu_main.npz', allow_pickle=True)
# mn_main, times_main, ps_main = res_main['mn'], res_main['times'], res_main['ps']
#
# plt.semilogy(mn_pr, times_pr, label='pr')
# plt.semilogy(mn_main, times_main, label='main')
# plt.xlabel('m = n')
# plt.ylabel('Execution Time (s)')
# plt.legend()
# plt.show()
#
# mns = set(mn_pr).intersection(set(mn_main))
# for mn in mns:
#     psi_pr = ps_pr[()][mn]
#     psi_main = ps_main[()][mn]
#     np.testing.assert_allclose(psi_pr, psi_main, rtol=5e-14)
```

</details>