@AtsushiSakai 2024-03-24
I believe this PR has been reviewed for a sufficient amount of time, so I will merge it. Thank you for the fantastic performance improvement!! @mdhaber